Updated AMI search instructions
The picture and instruction provided depicts your experience as the developers. As the user, we must first switch to 'Public images' prior to searching for said AMI ID. Following the current instructions and pictures, the user would obtain no results from their search. 

I only updated the documentation's steps and not the pictures. If you would like, I could do that as well.